Prepared logic to wipeout data. Repaired account deletion. Moved password validation logic to separate method. (#135)

+++ Application/Services/AccountService.cs
@@ -1,4 +1,5 @@
 ﻿using Application.Dtos.Accounts;
+using Application.Dtos.Auth;
 using Application.Interfaces;
 using AutoMapper;
 using Domain.Exceptions;
@@ -80,17 +81,25 @@ namespace Application.Services
 
         public async Task DeleteAccountAsync(int accountId, DeleteAccountDto deleteAccountDto, CancellationToken cancellationToken = default)
         {
-            var account = await _unitOfWork.Accounts.GetByIdAsync(accountId, cancellationToken).ConfigureAwait(false)
+            var accountForAuth = await _unitOfWork.Accounts.GetByIdAsync(accountId, cancellationToken).ConfigureAwait(false)
                 ?? throw new NotFoundException($"Account with ID: {accountId} not found");
 
-            var result = _passwordHasher.VerifyHashedPassword(account, account.HashPassword, deleteAccountDto.Password);
+            var result = _passwordHasher.VerifyHashedPassword(accountForAuth, accountForAuth.HashPassword, deleteAccountDto.Password);
             if (result != PasswordVerificationResult.Success)
                 throw new UnauthorizedException("Invalid password");
 
+            // First fetch and delete labels to avoid db conflict. Needs to be done separately because doing so in one operation violates non nullability on AccountId
             var labelsToDelete = await _unitOfWork.QuestLabels.GetUserLabelsAsync(accountId, false, cancellationToken).ConfigureAwait(false);
+            if (labelsToDelete.Any())
+            {
+                _unitOfWork.QuestLabels.DeleteRange(labelsToDelete);
+                await _unitOfWork.SaveChangesAsync(cancellationToken).ConfigureAwait(false);
+            }
 
-            _unitOfWork.Accounts.Delete(account);
-            _unitOfWork.QuestLabels.DeleteRange(labelsToDelete);
+            var accountToDelete = await _unitOfWork.Accounts.GetByIdAsync(accountId, cancellationToken).ConfigureAwait(false)
+                ?? throw new NotFoundException($"Account with ID: {accountId} not found after label deletion.");
+
+            _unitOfWork.Accounts.Delete(accountToDelete);
 
             await _unitOfWork.SaveChangesAsync(cancellationToken).ConfigureAwait(false);
         }
@@ -118,6 +127,28 @@ namespace Application.Services
                 _logger.LogInformation("Updated timezone for user {UserId} to {TimeZone}.", accountId, normalizedTimeZone);
                 await _unitOfWork.SaveChangesAsync(cancellationToken).ConfigureAwait(false);
             }
+        }
+
+        public async Task WipeoutAccountDataAsync(PasswordConfirmationDto authRequestDto, CancellationToken cancellationToken = default)
+        {
+            var account = await _unitOfWork.Accounts.GetAccountToWipeoutDataAsync(authRequestDto.AccountId, cancellationToken).ConfigureAwait(false)
+                ?? throw new NotFoundException($"Account with ID {authRequestDto.AccountId} was not found");
+
+            var result = _passwordHasher.VerifyHashedPassword(account, account.HashPassword, authRequestDto.Password);
+            if (result != PasswordVerificationResult.Success)
+                throw new UnauthorizedException("Invalid password");
+
+            // Clear profile information
+            account.Profile.WipeoutData();
+            // Clear quests
+            account.Quests.Clear();
+            // Clear labels
+            var labelsToDelete = account.Labels;
+            _unitOfWork.QuestLabels.DeleteRange(labelsToDelete);
+
+            var affectedRows = await _unitOfWork.SaveChangesAsync(cancellationToken).ConfigureAwait(false);
+
+            _logger.LogDebug("Affected rows after wiping out account data: {AffectedRows}", affectedRows);
         }
     }
 }
